+++ src/include/sll/io.h
@@ -90,4 +90,17 @@ __SLL_EXTERNAL void sll_write_node(sll_file_t* wf,const sll_node_t* o);
 
 
 
+/**
+ * \flags func
+ * \name sll_write_sll_code
+ * \group io
+ * \desc Docs!
+ * \arg const sll_compilation_data_t* c_dt
+ * \arg const sll_internal_function_table_t* i_ft
+ * \arg sll_file_t* wf
+ */
+__SLL_EXTERNAL void sll_write_sll_code(const sll_compilation_data_t* c_dt,const sll_internal_function_table_t* i_ft,sll_file_t* wf);
+
+
+
 #endif