+++ src/sll/file.c
@@ -313,12 +313,13 @@ __SLL_EXTERNAL sll_size_t sll_file_write(sll_file_t* f,const void* p,sll_size_t 
 	sll_copy_data(p,i,f->_w_bf+f->_w_bf_off);
 	sll_platform_file_write(f->dt.fl.fd,f->_w_bf,FILE_BUFFER_SIZE);
 	const sll_char_t* v=((const sll_char_t*)p)+i;
-	sll_size_t o=i;
+	sll_size_t o=sz;
 	sz-=i;
 	i=sz&(FILE_BUFFER_SIZE-1);
-	if (sz-i){
-		sll_platform_file_write(f->dt.fl.fd,v,sz-i);
-		v+=sz-i;
+	sz-=i;
+	if (sz){
+		sll_platform_file_write(f->dt.fl.fd,v,sz);
+		v+=sz;
 	}
 	if (i){
 		sll_copy_data(v,i,f->_w_bf);
@@ -352,6 +353,57 @@ __SLL_EXTERNAL sll_bool_t sll_file_write_char(sll_file_t* f,sll_char_t c){
 		f->_w_bf_off=0;
 	}
 	return 1;
+}
+
+
+__SLL_EXTERNAL sll_size_t sll_file_write_char_count(sll_file_t* f,sll_char_t c,sll_size_t n){
+	if (!(f->f&SLL_FILE_FLAG_WRITE)||!n){
+		return 0;
+	}
+	if (f->f&FILE_FLAG_MEMORY){
+		if (f->_off+n>=f->dt.mm.sz){
+			n=f->dt.mm.sz-f->_off;
+		}
+		sll_set_memory(((sll_char_t*)(f->dt.mm.p))+f->_off,n,c);
+		f->_off+=n;
+		return n;
+	}
+	if (f->f&SLL_FILE_FLAG_NO_BUFFER){
+		sll_size_t o=0;
+		while (n){
+			o+=sll_platform_file_write(f->dt.fl.fd,&c,sizeof(sll_char_t));
+			n--;
+		}
+		return o;
+	}
+	if (n+f->_w_bf_off<=FILE_BUFFER_SIZE){
+		sll_set_memory(f->_w_bf+f->_w_bf_off,n,c);
+		f->_w_bf_off+=n;
+		if (f->_w_bf_off==FILE_BUFFER_SIZE){
+			sll_platform_file_write(f->dt.fl.fd,f->_w_bf,FILE_BUFFER_SIZE);
+			f->_w_bf_off=0;
+		}
+		return n;
+	}
+	sll_size_t i=FILE_BUFFER_SIZE-f->_w_bf_off;
+	sll_set_memory(f->_w_bf+f->_w_bf_off,i,c);
+	sll_platform_file_write(f->dt.fl.fd,f->_w_bf,FILE_BUFFER_SIZE);
+	sll_size_t o=n;
+	n-=i;
+	i=n&(FILE_BUFFER_SIZE-1);
+	n-=i;
+	if (n){
+		sll_set_memory(f->_w_bf,FILE_BUFFER_SIZE,c);
+		while (n){
+			sll_platform_file_write(f->dt.fl.fd,f->_w_bf,FILE_BUFFER_SIZE);
+			n-=FILE_BUFFER_SIZE;
+		}
+	}
+	if (i){
+		sll_set_memory(f->_w_bf,i,c);
+	}
+	f->_w_bf_off=i;
+	return o;
 }